requirements.txt
certifi==2020.11.8
chardet==3.0.4
idna==2.10
progressbar2==3.53.1
python-utils==2.4.0
requests==2.25.0
six==1.15.0
urllib3==1.26.2
